probe-evasion/scripts/build_concept_dataset.py
"""Combine and validate concept training datasets from multiple sources."""

import argparse
import random
from collections import Counter
from pathlib import Path
from typing import Dict, List, Optional

import yaml


def load_yaml_examples(path: str) -> List[Dict]:
    """
    Load examples from a YAML file.

    Args:
        path: Path to YAML file with 'examples' key.

    Returns:
        List of example dicts.
    """
    with open(path, "r") as f:
        data = yaml.safe_load(f)
    return data.get("examples", [])


def validate_example(example: Dict) -> List[str]:
    """
    Validate a single example for required fields and consistency.

    Args:
        example: Dict with 'text', 'label', 'subcategory'.

    Returns:
        List of validation error strings (empty if valid).
    """
    errors = []
    if "text" not in example or not isinstance(example.get("text"), str):
        errors.append("Missing or invalid 'text' field")
    if "label" not in example or example.get("label") not in (0, 1):
        errors.append("Missing or invalid 'label' field (must be 0 or 1)")
    if "subcategory" not in example or not isinstance(example.get("subcategory"), str):
        errors.append("Missing or invalid 'subcategory' field")
    return errors


def check_class_balance(examples: List[Dict]) -> Dict:
    """
    Analyze class balance in dataset.

    Args:
        examples: List of example dicts.

    Returns:
        Dict with counts and balance metrics.
    """
    counts = Counter(ex["label"] for ex in examples)
    total = len(examples)
    return {
        "total": total,
        "positive": counts.get(1, 0),
        "negative": counts.get(0, 0),
        "positive_ratio": counts.get(1, 0) / total if total > 0 else 0,
        "negative_ratio": counts.get(0, 0) / total if total > 0 else 0,
    }


def check_subcategory_coverage(examples: List[Dict]) -> Dict:
    """
    Analyze subcategory distribution.

    Args:
        examples: List of example dicts.

    Returns:
        Dict mapping subcategory to count.
    """
    raise NotImplementedError("TODO")


def deduplicate(examples: List[Dict], threshold: float = 0.9) -> List[Dict]:
    """
    Remove near-duplicate examples.

    Uses simple text similarity to identify duplicates.

    Args:
        examples: List of example dicts.
        threshold: Similarity threshold for considering duplicates.

    Returns:
        Deduplicated list.
    """
    raise NotImplementedError("TODO")


def filter_by_length(
    examples: List[Dict],
    min_chars: int = 50,
    max_chars: int = 500
) -> List[Dict]:
    """
    Filter examples by text length.

    Args:
        examples: List of example dicts.
        min_chars: Minimum character count.
        max_chars: Maximum character count.

    Returns:
        Filtered list.
    """
    raise NotImplementedError("TODO")


def combine_sources(
    sources: List[str],
    target_positive: int = 500,
    target_negative: int = 500,
    seed: int = 42
) -> List[Dict]:
    """
    Combine examples from multiple source files.

    Balances the dataset and shuffles.

    Args:
        sources: List of paths to YAML files.
        target_positive: Target number of positive examples.
        target_negative: Target number of negative examples.
        seed: Random seed.

    Returns:
        Combined and balanced list of examples.
    """
    raise NotImplementedError("TODO")


def split_dataset(
    examples: List[Dict],
    train_ratio: float = 0.8,
    val_ratio: float = 0.1,
    test_ratio: float = 0.1,
    seed: int = 42
) -> Dict[str, List[Dict]]:
    """
    Split dataset into train/val/test sets.

    Maintains class balance in each split via stratification.

    Args:
        examples: List of example dicts.
        train_ratio: Fraction for training.
        val_ratio: Fraction for validation.
        test_ratio: Fraction for testing.
        seed: Random seed.

    Returns:
        Dict with 'train', 'val', 'test' keys.
    """
    from sklearn.model_selection import train_test_split

    labels = [ex["label"] for ex in examples]

    # First split: train vs (val + test)
    train_examples, remaining = train_test_split(
        examples,
        test_size=(val_ratio + test_ratio),
        random_state=seed,
        stratify=labels,
    )

    # Second split: val vs test
    remaining_labels = [ex["label"] for ex in remaining]
    relative_test_ratio = test_ratio / (val_ratio + test_ratio)
    val_examples, test_examples = train_test_split(
        remaining,
        test_size=relative_test_ratio,
        random_state=seed,
        stratify=remaining_labels,
    )

    return {
        "train": train_examples,
        "val": val_examples,
        "test": test_examples,
    }


def save_dataset(examples: List[Dict], path: str) -> None:
    """
    Save examples to YAML file.

    Args:
        examples: List of example dicts.
        path: Output file path.
    """
    raise NotImplementedError("TODO")


def print_dataset_stats(examples: List[Dict]) -> None:
    """
    Print summary statistics for a dataset.

    Args:
        examples: List of example dicts.
    """
    raise NotImplementedError("TODO")


def main():
    """Main entry point for dataset building."""
    parser = argparse.ArgumentParser(
        description="Combine and validate concept training datasets"
    )
    parser.add_argument(
        "--sources",
        type=str,
        nargs="+",
        required=True,
        help="Paths to source YAML files"
    )
    parser.add_argument(
        "--output-dir",
        type=str,
        default="data/concepts/trees",
        help="Output directory for final datasets"
    )
    parser.add_argument(
        "--target-positive",
        type=int,
        default=500,
        help="Target positive examples"
    )
    parser.add_argument(
        "--target-negative",
        type=int,
        default=500,
        help="Target negative examples"
    )
    parser.add_argument(
        "--seed",
        type=int,
        default=42,
        help="Random seed"
    )
    parser.add_argument(
        "--validate-only",
        action="store_true",
        help="Only validate, don't build"
    )

    args = parser.parse_args()

    raise NotImplementedError("TODO: Implement dataset building pipeline")


if __name__ == "__main__":
    main()
