probe-evasion/requirements.txt
torch
transformers
numpy
pandas
matplotlib
seaborn
pyyaml
tqdm
requests
wikipedia-api
bitsandbytes
accelerate
scikit-learn
